+++ src/DMS.Application/PriceLists/Dto/AssignPriceListDto.cs
@@ -1,0 +1,12 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace DMS.PriceLists.Dto;
+
+public class AssignPriceListDto
+{
+    [Required]
+    public int CustomerId { get; set; }
+
+    [Required]
+    public int PriceListId { get; set; }
+}